+++ alloff.php
@@ -1,18 +1,4 @@
 ﻿
-
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN"
-         "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml">
-<head>
-  <title>Lights</title>
-  <meta name="viewport" content="width=device-width; initial-scale=1.0; maximum-scale=1.0; user-scalable=0;"/>
-  <link rel="apple-touch-icon" href="msc_icon.png" />
-  <style type="text/css" media="screen">@import "iui/iui.css";</style>
-  <script type="application/x-javascript" src="iui/iui.js"></script>
-
-</head>
-<body>
-
 <?php
 include 'functions.php';
 
@@ -38,17 +24,9 @@ if($id == 'O')
   oneOff(5);
   oneOff(9);
   oneOff(16);
-  oneOff(3);
-  oneOff(19);
+  oneOff(36);
   $output = `/usr/bin/env curl -s http://max.kenkl.org/lights/brdloff.php`;
 }
 
 if(strpos($_SERVER['HTTP_USER_AGENT'], 'WebKit')) header('Location: ' . $_SERVER['alloff.php']);
 ?>
-</body>
-</html>
-
-
-
-
-