+++ brfull.php
@@ -1,24 +1,10 @@
 ﻿
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN"
-         "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml">
-<head>
-  <title>Lights</title>
-  <meta name="viewport" content="width=device-width; initial-scale=1.0; maximum-scale=1.0; user-scalable=0;"/>
-  <link rel="apple-touch-icon" href="msc_icon.png" />
-  <style type="text/css" media="screen">@import "iui/iui.css";</style>
-  <script type="application/x-javascript" src="iui/iui.js"></script>
-
-</head>
-<body>
-
 <?php
 require 'functions.php';
 
 oneOnBright(5);
 oneOnBright(16);
-oneOnBright(3);
-oneOnBright(19);
+oneOnBright(36);
 oneOn(9);
 
 $output = `/usr/bin/curl http://max.kenkl.org/lights/brdlonfullwarm.php`;
@@ -26,6 +12,3 @@ $output = `/usr/bin/curl http://max.kenkl.org/lights/brdlonfullwarm.php`;
 if(strpos($_SERVER['HTTP_USER_AGENT'], 'WebKit')) header('Location: ' . $_SERVER['brfull.php']);
 
 ?>
-</body>
-</html>
-